cityLBM用の調整

+++ IS_DaemonAndSampler/InSituLib/unstruct/kvs_wrapper.h
@@ -138,7 +138,8 @@ extern "C" {
         int ncells;
         int  MPIprocess_per_ensemble; 
         int  ensemble_number; 
-        int  repitetion; 
+        //int  repitetion; 
+        float  repitetion; 
     } Ensembleparameters;
